# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aqary Real Estate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/seller/activity:
    get:
      operationId: getSellerActivity
      tags: [properties]
      summary: Recent interactions on seller listings
      responses:
        "200":
          description: Recent activity for current seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerActivityResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all approved properties
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropertyItem"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create property listing (seller)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update property (seller)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyItem"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete property (seller)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PropertyItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        location:
          type: string
        area:
          type: number
        rooms:
          type: ["integer", "null"]
        propertyType:
          type: string
        features:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        sellerId:
          type: integer
        status:
          type: string
        views:
          type: integer
        saves:
          type: integer
        contacts:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - price
        - location
        - area
        - propertyType
        - features
        - sellerId
        - status
        - views
        - saves
        - contacts
        - createdAt
    CreatePropertyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        location:
          type: string
        area:
          type: number
        rooms:
          type: integer
        propertyType:
          type: string
          enum: [apartment, villa, commercial, land]
        features:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
          maxItems: 3
      required:
        - title
        - description
        - price
        - location
        - area
        - propertyType
    UpdatePropertyBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        location:
          type: string
        area:
          type: number
        rooms:
          type: integer
        propertyType:
          type: string
        features:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
          maxItems: 3
...
    SellerActivityEvent:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyTitle:
          type: string
        interactionType:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - propertyId
        - propertyTitle
        - interactionType
        - createdAt
    SellerActivityResponse:
      type: object
      properties:
        recentInteractions:
          type: array
          items:
            $ref: "#/components/schemas/SellerActivityEvent"
      required:
        - recentInteractions
...